LPS-153401 Move code related to Code Editor to Util class
